// src/thrift/bindings/ServiceRegistry_types.d.ts
//
// Autogenerated by Thrift Compiler (0.24.0)
//
// DO NOT EDIT UNLESS YOU ARE SURE THAT YOU KNOW WHAT YOU ARE DOING
//
import thrift = require('thrift');
import Thrift = thrift.Thrift;
import Int64 = require('node-int64');
import { v4 as uuid } from 'uuid';
type uuid = string;


declare enum Protocol {
  Binary = 0,
  Json = 1,
}

declare enum Transport {
  Socket = 0,
  Pipe = 1,
}

declare class ServiceLocation {
    public host: string;
    public port: number;
    public protocol: Protocol;
    public transport: Transport;

      read(input: object): void;
  write(output: object): void;
  constructor(args?: { host: string; port: number; protocol: Protocol; transport: Transport; });
  }

declare class ServiceException extends Thrift.TException {
    public message: string;
    public serviceName: string;
    public exceptionDescr: string;

      read(input: object): void;
  write(output: object): void;
  constructor(args?: { message: string; serviceName: string; exceptionDescr: string; });
  }

declare const IAR_SERVICE_REGISTRY_ENVVAR: string;

declare const IAR_CSPY_PIPE_PREFIX: string;

declare const SERVICE_REGISTRY_SERVICE: string;